Changed DataSourceController and repository to return ID of the inserted model

+++ DocumentDataAPI/DocumentDataAPI/Data/Repositories/IRepository.cs
@@ -4,7 +4,7 @@ public interface IRepository<TEntity>
     where TEntity : class
 {
     public Task<IEnumerable<TEntity>> GetAll();
-    public Task<int> Add(TEntity entity);
+    public Task<long> Add(TEntity entity);
     public Task<int> Delete(TEntity entity);
     public Task<int> Update(TEntity entity);
 }
